Reward system now accepts double (not int) as reward scalar to improve control of reward amount by individual tasks

+++ Behavioural Testing Unit (Android)/Mymou/app/src/main/java/mymou/task/backend/TaskInterface.java
@@ -2,6 +2,6 @@ package mymou.task.backend;
 
 public interface TaskInterface {
     void resetTimer_();
-    void trialEnded_(String outcome, int rew_scalar);
+    void trialEnded_(String outcome, double rew_scalar);
     void logEvent_(String event);
 }
